requirements.txt
fastapi
uvicorn
openai
faiss-cpu
pandas
python-dotenv
pyarrow